store sales plan data

+++ database/migrations/2022_08_22_074511_create_sales_plan_additional_costs_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('sales_plan_id')->constrained();
             $table->foreignId('additional_cost_id')->constrained();
-            $table->float('percentage')->default(0);
-            $table->float('amount')->default(0);
+            $table->double('percentage')->default(0);
+            $table->double('amount')->default(0);
             $table->timestamps();
         });
     }
